+++ src/main/java/com/box/sdk/BoxFolder.java
@@ -370,7 +370,7 @@ public class BoxFolder extends BoxItem implements Iterable<BoxItem.Info> {
      * @param recursive true to recursively delete this folder's contents; otherwise false.
      */
     public void delete(boolean recursive) {
-        URL url = DELETE_FOLDER_URL.build(this.getAPI().getBaseURL(), this.getID(), recursive);
+        URL url = DELETE_FOLDER_URL.buildAlpha(this.getAPI().getBaseURL(), this.getID(), recursive);
         BoxAPIRequest request = new BoxAPIRequest(this.getAPI(), url, "DELETE");
         BoxAPIResponse response = request.send();
         response.disconnect();
@@ -849,7 +849,7 @@ public class BoxFolder extends BoxItem implements Iterable<BoxItem.Info> {
      * @return the metadata returned from the server.
      */
     public Metadata createMetadata(String templateName, String scope, Metadata metadata) {
-        URL url = METADATA_URL_TEMPLATE.build(this.getAPI().getBaseURL(), this.getID(), scope, templateName);
+        URL url = METADATA_URL_TEMPLATE.buildAlpha(this.getAPI().getBaseURL(), this.getID(), scope, templateName);
         BoxAPIRequest request = new BoxAPIRequest(this.getAPI(), url, "POST");
         request.addHeader("Content-Type", "application/json");
         request.setBody(metadata.toString());
@@ -876,7 +876,7 @@ public class BoxFolder extends BoxItem implements Iterable<BoxItem.Info> {
                 for (JsonValue value : metadata.getOperations()) {
                     if (value.asObject().get("value").isNumber()) {
                         metadataToUpdate.add(value.asObject().get("path").asString(),
-                                value.asObject().get("value").asFloat());
+                                value.asObject().get("value").asDouble());
                     } else if (value.asObject().get("value").isString()) {
                         metadataToUpdate.add(value.asObject().get("path").asString(),
                                 value.asObject().get("value").asString());
@@ -925,7 +925,7 @@ public class BoxFolder extends BoxItem implements Iterable<BoxItem.Info> {
      * @return the metadata returned from the server.
      */
     public Metadata getMetadata(String templateName, String scope) {
-        URL url = METADATA_URL_TEMPLATE.build(this.getAPI().getBaseURL(), this.getID(), scope, templateName);
+        URL url = METADATA_URL_TEMPLATE.buildAlpha(this.getAPI().getBaseURL(), this.getID(), scope, templateName);
         BoxAPIRequest request = new BoxAPIRequest(this.getAPI(), url, "GET");
         BoxJSONResponse response = (BoxJSONResponse) request.send();
         return new Metadata(JsonObject.readFrom(response.getJSON()));
@@ -938,7 +938,7 @@ public class BoxFolder extends BoxItem implements Iterable<BoxItem.Info> {
      * @return the metadata returned from the server.
      */
     public Metadata updateMetadata(Metadata metadata) {
-        URL url = METADATA_URL_TEMPLATE.build(this.getAPI().getBaseURL(), this.getID(), metadata.getScope(),
+        URL url = METADATA_URL_TEMPLATE.buildAlpha(this.getAPI().getBaseURL(), this.getID(), metadata.getScope(),
                 metadata.getTemplateName());
         BoxAPIRequest request = new BoxAPIRequest(this.getAPI(), url, "PUT");
         request.addHeader("Content-Type", "application/json-patch+json");
@@ -971,7 +971,7 @@ public class BoxFolder extends BoxItem implements Iterable<BoxItem.Info> {
      * @param scope        the scope of the template (usually "global" or "enterprise").
      */
     public void deleteMetadata(String templateName, String scope) {
-        URL url = METADATA_URL_TEMPLATE.build(this.getAPI().getBaseURL(), this.getID(), scope, templateName);
+        URL url = METADATA_URL_TEMPLATE.buildAlpha(this.getAPI().getBaseURL(), this.getID(), scope, templateName);
         BoxAPIRequest request = new BoxAPIRequest(this.getAPI(), url, "DELETE");
         BoxAPIResponse response = request.send();
         response.disconnect();
@@ -1211,10 +1211,11 @@ public class BoxFolder extends BoxItem implements Iterable<BoxItem.Info> {
         private boolean canNonOwnersInvite;
         private boolean isWatermarked;
         private boolean isCollaborationRestrictedToEnterprise;
-        private Boolean isExternallyOwned;
+        private boolean isExternallyOwned;
         private Map<String, Map<String, Metadata>> metadataMap;
         private List<String> allowedSharedLinkAccessLevels;
         private List<String> allowedInviteeRoles;
+        private BoxClassification classification;
 
         /**
          * Constructs an empty Info object.
@@ -1398,6 +1399,14 @@ public class BoxFolder extends BoxItem implements Iterable<BoxItem.Info> {
             return this.isExternallyOwned;
         }
 
+        /**
+         * Gets the metadata classification type of this folder.
+         * @return the metadata classification type of this folder.
+         */
+        public BoxClassification getClassification() {
+            return this.classification;
+        }
+
         @Override
         public BoxFolder getResource() {
             return BoxFolder.this;
@@ -1442,6 +1451,12 @@ public class BoxFolder extends BoxItem implements Iterable<BoxItem.Info> {
                 } else if (memberName.equals("metadata")) {
                     JsonObject jsonObject = value.asObject();
                     this.metadataMap = Parsers.parseAndPopulateMetadataMap(jsonObject);
+                } else if (memberName.equals("classification")) {
+                    if (value.isNull()) {
+                        this.classification = null;
+                    } else {
+                        this.classification = new BoxClassification(value.asObject());
+                    }
                 }
             } catch (Exception e) {
                 throw new BoxDeserializationException(memberName, value.toString(), e);