+++ src/main/java/com/box/sdk/BoxGlobalSettings.java
@@ -6,7 +6,7 @@ package com.box.sdk;
 public final class BoxGlobalSettings {
     private static int connectTimeout = 0;
     private static int readTimeout = 0;
-    private static int maxRequestAttempts = BoxAPIConnection.DEFAULT_MAX_ATTEMPTS;
+    private static int maxRetryAttempts = BoxAPIConnection.DEFAULT_MAX_RETRIES;
 
     private BoxGlobalSettings() {
     }
@@ -44,18 +44,44 @@ public final class BoxGlobalSettings {
     }
 
     /**
-     * Returns the global maximum number of request attempts.
+     * Returns the global total maximum number of times an API request will be tried when error responses
+     * are received.
      * @return max number of request attempts
+     * @deprecated getMaxRetryAttempts is preferred because it more clearly gets the number
+     * of times a request should be retried after an error response is received.
      */
+    @Deprecated
     public static int getMaxRequestAttempts() {
-        return maxRequestAttempts;
+        return maxRetryAttempts + 1;
     }
 
     /**
-     * Sets the global default maximum number of request attempts.
-     * @param maxRequestAttempts maximum number of request attempts
+     * Sets the global total maximum number of times an API request will be tried when error responses
+     * are received.
+     * @param attempts maximum number of request attempts
+     * @deprecated setMaxRetryAttempts is preferred because it more clearly sets the number
+     * of times a request should be retried after an error response is received.
      */
-    public static void setMaxRequestAttempts(int maxRequestAttempts) {
-        BoxGlobalSettings.maxRequestAttempts = maxRequestAttempts;
+    @Deprecated
+    public static void setMaxRequestAttempts(int attempts) {
+        BoxGlobalSettings.maxRetryAttempts = attempts - 1;
+    }
+
+    /**
+     * Returns the global maximum number of times an API request will be retried after an error response
+     * is received.
+     * @return max number of request attempts
+     */
+    public static int getMaxRetryAttempts() {
+        return maxRetryAttempts;
+    }
+
+    /**
+     * Sets the global maximum number of times an API request will be retried after an error response
+     * is received.
+     * @param attempts maximum number of request attempts
+     */
+    public static void setMaxRetryAttempts(int attempts) {
+        BoxGlobalSettings.maxRetryAttempts = attempts;
     }
 }