+++ src/main/java/com/box/sdk/Metadata.java
@@ -178,8 +178,23 @@ public class Metadata {
      * @param path the path that designates the key. Must be prefixed with a "/".
      * @param value the value.
      * @return this metadata object.
+     * @deprecated add(String, double) is preferred as it avoids errors when converting a
+     * float to the underlying data type used by Metadata (double)
      */
+    @Deprecated
     public Metadata add(String path, float value) {
+        this.values.add(this.pathToProperty(path), value);
+        this.addOp("add", path, value);
+        return this;
+    }
+
+    /**
+     * Adds a new metadata value.
+     * @param path the path that designates the key. Must be prefixed with a "/".
+     * @param value the value.
+     * @return this metadata object.
+     */
+    public Metadata add(String path, double value) {
         this.values.add(this.pathToProperty(path), value);
         this.addOp("add", path, value);
         return this;
@@ -220,6 +235,18 @@ public class Metadata {
      * @return this metadata object.
      */
     public Metadata replace(String path, float value) {
+        this.values.set(this.pathToProperty(path), value);
+        this.addOp("replace", path, value);
+        return this;
+    }
+
+    /**
+     * Replaces an existing metadata value.
+     * @param path the path that designates the key. Must be prefixed with a "/".
+     * @param value the value.
+     * @return this metadata object.
+     */
+    public Metadata replace(String path, double value) {
         this.values.set(this.pathToProperty(path), value);
         this.addOp("replace", path, value);
         return this;
@@ -316,12 +343,23 @@ public class Metadata {
     }
 
     /**
-     * Get a value from a float metadata field.
+     * Get a value from a double metadata field.
+     * @param path the key path in the metadata object.  Must be prefixed with a "/".
+     * @return the metadata value as a double floating point number.
+     * @deprecated getDouble() is preferred as it more clearly describes the return type (double)
+     */
+    @Deprecated
+    public double getFloat(String path) {
+        // @NOTE(mwiller) 2018-02-05: JS number are all 64-bit floating point, so double is the correct type to use here
+        return this.getValue(path).asDouble();
+    }
+
+    /**
+     * Get a value from a double metadata field.
      * @param path the key path in the metadata object.  Must be prefixed with a "/".
      * @return the metadata value as a floating point number.
      */
-    public double getFloat(String path) {
-        // @NOTE(mwiller) 2018-02-05: JS number are all 64-bit floating point, so double is the correct type to use here
+    public double getDouble(String path) {
         return this.getValue(path).asDouble();
     }
 
@@ -452,6 +490,22 @@ public class Metadata {
                 .add("value", value));
     }
 
+    /**
+     * Adds a patch operation.
+     * @param op the operation type. Must be add, replace, remove, or test.
+     * @param path the path that designates the key. Must be prefixed with a "/".
+     * @param value the value to be set.
+     */
+    private void addOp(String op, String path, double value) {
+        if (this.operations == null) {
+            this.operations = new JsonArray();
+        }
+
+        this.operations.add(new JsonObject()
+                .add("op", op)
+                .add("path", path)
+                .add("value", value));
+    }
     /**
      * Adds a new patch operation for array values.
      * @param op the operation type. Must be add, replace, remove, or test.